flattens the project file structure
Rather than server/ and client/ directories, I flattened the directory
so that server and client are a bit more mixed up. The initial reason
for this is that nesting feathers code inside a `server/` directory was
causing issues with the generator commands, since they expect a very
definite file structure. As I un-nested the feathers files, I had to
decide what to do with the "client" files. Ultimately, I decided that
un-nesting them from the `client/` directory, to live side-by-side with
the server files, was the right move. This exposes the entire structure
of the application at once, and since we're a Universal app with both
client and server rendering, it makes sense to me that all the code is
part of one application. Further, since Lauchpad makes decisions about
frontend frameworks (React, Redux) and their setup, making it appear to
be part of the core application and not "pluggable" benefits the way we
should be thinking about our applications as we build them: as a way of
rendering and navigating a set of defined services.